pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapmerge"
version = "0.0.0"   # The effective version is shown from app_version
requires-python = ">=3.11"
dependencies = [
  "pywin32>=306",
  "PySide6>=6.6.0",
  "pillow>=10.0.0",
  "PyPDF2>=3.0.0",
  "docx2pdf>=0.1.8; platform_system == 'Windows'",
  "typer>=0.12.3",
  "pycryptodome>=3.20",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]